split ContentGenerator and TypingString classes into different files
